// src/components/ReportFormed/components/ReportDownloadFiles/types.ts
export enum EReportFormedTags {
  common = "common",
  expenses = "expenses",
  incomes = "incomes",
  others = "others",
}

export enum EReportFormedTagsRu {
  common = "reportFormedTags.keyDocuments",
  expenses = "reportFormedTags.confirmationOfExpenses",
  incomes = "reportFormedTags.incomeConfirmation",
  others = "reportFormedTags.otherDocuments",
}